Fix signed/unsigned warning

+++ src/Input.cpp
@@ -97,7 +97,7 @@ void Input::handleRead(const std::vector<char>& data)
                 reply.time = 0;
                 reply.zero = 0;
                 
-                for (int i = 0; i < sizeof(reply.randomBytes); ++i)
+                for (size_t i = 0; i < sizeof(reply.randomBytes); ++i)
                 {
                     reply.randomBytes[i] = static_cast<uint8_t>(_generator() % 255);
                 }
